iser: generating cmd batches using NOT_LAST bit
As iser accumulates received scsi commands on a list, scheduling an event which
processes the list, it is able to generate batches of commands submitted to scsi.

This patch uses NOT_LAST bit added to the cmd state to mark all commands in the
batch except the last one NOT_LAST=1.

It also makes sure that the last command in batch is a regular one:
read/write/sync_cache. Other types of commands may be executed synchronously
by the scsi engine, and if no commands arrive for some time, the backing store
which did not see the last command may stall submission of the previous commands.

Signed-off-by: Alexander Nezhinsky <alexandern@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>