MMC: Allow creating MMC LUNS that do not have a backingstore.
An MMC device without a backing store is emulated as an empty drive, with no media in it. TGTD will respond to all scsi i/o just like a scsi dved drive that does not have any media.

As soon as a backing store is added at a later time, the TGTD emulation will switch to instead emulate a DVD drive with either a blank disk in the tray, that can be written to (if the file is zero-length) or as a DVD drive with a DVD-ROM disk in the tray (if the backing store is a .iso image)

Being able to create DVD drives without any backing file is useful for when creating a jukebox. In that case it is most convenient to create the drive without any media loaded and then later use MTX to move media from astorage slot into the drive.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>